# docs/hr-service-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "CustomersController_findAll",
        "summary": "Listar clientes del almacén",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": "10"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nombre o documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de clientes"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "summary": "Registrar nuevo cliente",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente creado exitosamente"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/customers/pos-search": {
      "get": {
        "operationId": "CustomersController_searchForPos",
        "summary": "Búsqueda rápida de clientes para el POS (máx. 10 resultados)",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Nombre o número de documento",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de búsqueda"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomersController_findById",
        "summary": "Obtener cliente por ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "summary": "Actualizar datos del cliente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente actualizado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "summary": "Eliminar cliente (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del cliente",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente eliminado"
          },
          "404": {
            "description": "Cliente no encontrado"
          }
        },
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/vendors": {
      "get": {
        "operationId": "VendorsController_findAll",
        "summary": "Listar proveedores del almacén",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": "10"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nombre, teléfono o dirección",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de proveedores"
          }
        },
        "tags": [
          "Vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "VendorsController_create",
        "summary": "Registrar nuevo proveedor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proveedor creado exitosamente"
          }
        },
        "tags": [
          "Vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/vendors/{id}": {
      "get": {
        "operationId": "VendorsController_findById",
        "summary": "Obtener proveedor por ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del proveedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor encontrado"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "tags": [
          "Vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "VendorsController_update",
        "summary": "Actualizar datos del proveedor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del proveedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "tags": [
          "Vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "VendorsController_remove",
        "summary": "Eliminar proveedor (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del proveedor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Proveedor eliminado"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "tags": [
          "Vendors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attendance": {
      "post": {
        "operationId": "AttendanceController_record",
        "summary": "Registrar o actualizar asistencia (upsert por teamId + fecha)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asistencia registrada o actualizada"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "tags": [
          "Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attendance/daily": {
      "get": {
        "operationId": "AttendanceController_getDailySummary",
        "summary": "Resumen de asistencia del día para el almacén",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Fecha en formato YYYY-MM-DD",
            "schema": {
              "example": "2026-08-25",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de asistencia de todos los miembros"
          }
        },
        "tags": [
          "Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/attendance/monthly": {
      "get": {
        "operationId": "AttendanceController_getByMonth",
        "summary": "Asistencia mensual por colaborador o resumen por almacén",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Mes en formato YYYY-MM",
            "schema": {
              "example": "2026-08",
              "type": "string"
            }
          },
          {
            "name": "teamId",
            "required": false,
            "in": "query",
            "description": "UUID del colaborador; si se envía, retorna registros por fecha",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registros del mes o resumen por empleado"
          }
        },
        "tags": [
          "Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Novedades Maritex — HR Service",
    "description": "API de Recursos Humanos: gestión de equipos de trabajo, asistencia diaria, pagos de personal, clientes y proveedores del almacén.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpsertCustomerDto": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string",
            "example": "74935446"
          },
          "name": {
            "type": "string",
            "example": "Anali Saraeli"
          },
          "surname": {
            "type": "string",
            "example": "Trujillo Cardenas"
          }
        },
        "required": [
          "dni",
          "name",
          "surname"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string",
            "example": "74935446"
          },
          "name": {
            "type": "string",
            "example": "Anali Saraeli"
          },
          "surname": {
            "type": "string",
            "example": "Trujillo Cardenas"
          }
        }
      },
      "UpsertVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Distribuidora Norte S.A.C.",
            "maxLength": 255
          },
          "address": {
            "type": "string",
            "example": "Av. Industrial 123, Lima",
            "maxLength": 255
          },
          "local": {
            "type": "string",
            "example": "Galería Central",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "example": "999888777",
            "maxLength": 20
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateVendorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Distribuidora Norte S.A.C.",
            "maxLength": 255
          },
          "address": {
            "type": "string",
            "example": "Av. Industrial 123, Lima",
            "maxLength": 255
          },
          "local": {
            "type": "string",
            "example": "Galería Central",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "example": "999888777",
            "maxLength": 20
          }
        }
      },
      "RecordAttendanceDto": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-08-25"
          },
          "status": {
            "type": "string",
            "enum": [
              "PUNTUAL",
              "TOLERANCIA",
              "TARDE",
              "FALTA",
              "FALTA_INJUSTIFICADA",
              "DESCANSO",
              "VACACIONES",
              "RECUPERACION",
              "VALDEO"
            ]
          },
          "checkIn": {
            "type": "string",
            "example": "09:05",
            "description": "Hora de entrada (HH:mm)"
          },
          "checkOut": {
            "type": "string",
            "example": "18:00"
          },
          "delayMinutes": {
            "type": "number",
            "description": "Minutos de retraso",
            "minimum": 0
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "teamId",
          "date",
          "status"
        ]
      },
...
    }
  }
}